sửa thêm phần reset thông tin ký trong signData với action Nội bộ từ chối (trả lại)

+++ src/main/java/com/vsm/business/service/custom/processRequest/ProcessRequestCustomService.java
@@ -890,6 +890,9 @@ public class ProcessRequestCustomService {
             // Khi trả lại thành công -> revert lại các file tài liệu chính đã ẩn đi khi ký thành công
             this.signUtils.revertFilePrimary(requestDataId);
 
+            // trả lại -> reset lại thông tin trong signData + không xóa OTP
+            this.resetDataCustomer(requestDataId, false);
+
             return true;
         };
     }
@@ -1918,7 +1921,7 @@ public class ProcessRequestCustomService {
             this.signUtils.revertFilePrimary(requestDataId);
 
             // khách hàng trả lại -> reset lại thông tin trong signData
-            this.resetDataCustomer(requestDataId);
+            this.resetDataCustomer(requestDataId, true);
 
             return true;
         };
@@ -1927,9 +1930,10 @@ public class ProcessRequestCustomService {
     /**
      * Hàm thực hiện reset lại thông tin các object của khách hàng (xóa mã OTP, reset lại số lần xem, lần ký, lần tải, ... trong signData)
      * @param requestDataId : id của phiếu
+     * @param isCustomer    : cờ xem có phải khách hàng hay không. true: khách hàng | false: không phải (nếu là khách hàng thì xóa thông tin otp)
      */
     @Transactional
-    private void resetDataCustomer(Long requestDataId){
+    private void resetDataCustomer(Long requestDataId, boolean isCustomer){
         this.signDataRepository.findAllByRequestDataId(requestDataId).forEach(ele -> {
             ele.setNumberSign(0L);
             ele.setNumberView(0L);
@@ -1938,8 +1942,10 @@ public class ProcessRequestCustomService {
             ele.setStatus(null);
             this.signDataRepository.save(ele);
         });
-        List<Long> otpIds = this.otpRepository.findAllByRequestDataId(requestDataId).stream().map(ele -> ele.getId()).collect(Collectors.toList());
-        this.otpRepository.deleteAllById(otpIds);
+        if(isCustomer){
+            List<Long> otpIds = this.otpRepository.findAllByRequestDataId(requestDataId).stream().map(ele -> ele.getId()).collect(Collectors.toList());
+            this.otpRepository.deleteAllById(otpIds);
+        }
     }
